# Conflicts:
#	RockWeb/Blocks/Groups/GroupDetailLava.ascx.cs

+++ RockWeb/Blocks/Groups/GroupDetailLava.ascx.cs
@@ -345,6 +345,22 @@ namespace RockWeb.Blocks.Groups
             get { return this.GetAttributeValue( AttributeKey.EnableCommunicationPreference ).AsBooleanOrNull() ?? false; }
         }
 
+        /// <summary>
+        /// Gets a value indicating whether the current person may edit the group.
+        /// </summary>
+        private bool CanEdit
+        {
+            get { return IsAuthorizedForGroup( Authorization.EDIT ); }
+        }
+
+        /// <summary>
+        /// Gets a value indicating whether the current person may manage group members.
+        /// </summary>
+        private bool CanManageMembers
+        {
+            get { return CanEdit || IsAuthorizedForGroup( Authorization.MANAGE_MEMBERS ); }
+        }
+
         #endregion
 
         #region Base Control Methods
@@ -389,7 +405,7 @@ namespace RockWeb.Blocks.Groups
                 RockContext rockContext = RockApp.Current.CreateRockContext();
                 GroupMemberService groupMemberService = new GroupMemberService( rockContext );
 
-                var groupMember = groupMemberService.Get( this.CurrentGroupMemberId );
+                var groupMember = groupMemberService.Queryable().FirstOrDefault( m => m.Id == this.CurrentGroupMemberId && m.GroupId == _groupId );
 
                 if ( groupMember == null )
                 {
@@ -466,7 +482,7 @@ namespace RockWeb.Blocks.Groups
 
             Group group = groupService.Get( _groupId );
 
-            if ( group != null && group.IsAuthorized( Authorization.EDIT, CurrentPerson ) )
+            if ( group != null && CanEdit )
             {
                 group.Name = tbName.Text;
                 group.Description = tbDescription.Text;
@@ -617,12 +633,17 @@ namespace RockWeb.Blocks.Groups
         /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
         protected void btnSaveGroupMember_Click( object sender, EventArgs e )
         {
+            if ( !CanManageMembers )
+            {
+                return;
+            }
+
             var rockContext = RockApp.Current.CreateRockContext();
             GroupMemberService groupMemberService = new GroupMemberService( rockContext );
 
             GroupTypeRole role = new GroupTypeRoleService( rockContext ).Get( ddlGroupRole.SelectedValueAsInt() ?? 0 );
 
-            var groupMember = groupMemberService.Get( this.CurrentGroupMemberId );
+            var groupMember = groupMemberService.Queryable().FirstOrDefault( m => m.Id == this.CurrentGroupMemberId && m.GroupId == _groupId );
             if ( groupMember == null )
             {
                 groupMember = new GroupMember { Id = 0 };
@@ -735,12 +756,17 @@ namespace RockWeb.Blocks.Groups
         {
             mdConfirmDelete.Hide();
 
+            if ( !CanManageMembers )
+            {
+                return;
+            }
+
             if ( GetAttributeValue( AttributeKey.AllowGroupMemberDelete ).AsBoolean() )
             {
                 RockContext rockContext = RockApp.Current.CreateRockContext();
                 GroupMemberService groupMemberService = new GroupMemberService( rockContext );
 
-                var groupMember = groupMemberService.Get( this.CurrentGroupMemberId );
+                var groupMember = groupMemberService.Queryable().FirstOrDefault( m => m.Id == this.CurrentGroupMemberId && m.GroupId == _groupId );
                 if ( groupMember != null )
                 {
                     groupMemberService.Delete( groupMember );
@@ -841,6 +867,11 @@ namespace RockWeb.Blocks.Groups
 
                 if ( eventArgs.Length == 2 )
                 {
+                    if ( !CanManageMembers )
+                    {
+                        return;
+                    }
+
                     string action = eventArgs[0];
                     string parameters = eventArgs[1];
 
@@ -901,6 +932,31 @@ namespace RockWeb.Blocks.Groups
                 pnlEditGroupMember.Visible = false;
                 DisplayViewGroup();
             }
+        }
+
+        /// <summary>
+        /// Determines whether the current person is authorized for the security action on the group.
+        /// Security role groups allow only the view action through this block.
+        /// </summary>
+        /// <param name="action">The security action, such as <see cref="Authorization.EDIT"/>.</param>
+        /// <returns><c>true</c> if the current person is authorized; otherwise, <c>false</c>.</returns>
+        private bool IsAuthorizedForGroup( string action )
+        {
+            var group = GroupCache.Get( _groupId );
+
+            if ( group == null )
+            {
+                return false;
+            }
+
+            var isSecurityRole = group.IsSecurityRole || group.GroupTypeId == GroupTypeCache.GetSecurityRoleGroupType()?.Id;
+
+            if ( isSecurityRole && action != Authorization.VIEW )
+            {
+                return false;
+            }
+
+            return group.IsAuthorized( action, CurrentPerson );
         }
 
         /// <summary>
@@ -975,8 +1031,8 @@ namespace RockWeb.Blocks.Groups
                 // add collection of allowed security actions
                 Dictionary<string, object> securityActions = new Dictionary<string, object>();
                 securityActions.Add( "View", group != null && group.IsAuthorized( Authorization.VIEW, CurrentPerson ) );
-                securityActions.Add( "ManageMembers", group != null && group.IsAuthorized( Authorization.MANAGE_MEMBERS, CurrentPerson ) );
-                securityActions.Add( "Edit", group != null && group.IsAuthorized( Authorization.EDIT, CurrentPerson ) );
+                securityActions.Add( "ManageMembers", CanManageMembers );
+                securityActions.Add( "Edit", CanEdit );
                 securityActions.Add( "Administrate", group != null && group.IsAuthorized( Authorization.ADMINISTRATE, CurrentPerson ) );
                 mergeFields.Add( "AllowedActions", securityActions );
 
@@ -1312,7 +1368,7 @@ namespace RockWeb.Blocks.Groups
             RockContext rockContext = RockApp.Current.CreateRockContext();
             GroupMemberService groupMemberService = new GroupMemberService( rockContext );
 
-            var groupMember = groupMemberService.Get( groupMemberId );
+            var groupMember = groupMemberService.Queryable().FirstOrDefault( m => m.Id == groupMemberId && m.GroupId == _groupId );
 
             if ( groupMember == null )
             {
@@ -1374,7 +1430,7 @@ namespace RockWeb.Blocks.Groups
             RockContext rockContext = RockApp.Current.CreateRockContext();
             GroupMemberService groupMemberService = new GroupMemberService( rockContext );
 
-            var groupMember = groupMemberService.Get( groupMemberId );
+            var groupMember = groupMemberService.Queryable().FirstOrDefault( m => m.Id == groupMemberId && m.GroupId == _groupId );
             if ( groupMember != null )
             {
                 // persist the group member id for use in partial postbacks
